Segmenting cfg and block env impl from PayloadBuilderAttributes
