+++ bin/reth/Cargo.toml
@@ -164,7 +164,6 @@ optimism = [
     "reth-node-ethereum/optimism",
     "dep:reth-node-optimism",
     "reth-node-core/optimism",
-    "reth-node-api/optimism",
 ]
 
 # no-op feature flag for switching between the `optimism` and default functionality in CI matrices
