+++ crates/node-api/Cargo.toml
@@ -19,6 +19,3 @@ thiserror.workspace = true
 
 # io
 serde.workspace = true
-
-[features]
-optimism = []
