+++ crates/node-core/Cargo.toml
@@ -122,7 +122,6 @@ optimism = [
     "reth-blockchain-tree/optimism",
     "reth-beacon-consensus/optimism",
     "reth-optimism-payload-builder/optimism",
-    "reth-node-api/optimism",
 ]
 
 jemalloc = ["dep:jemalloc-ctl"]
