+++ crates/node-optimism/Cargo.toml
@@ -35,7 +35,6 @@ reth-db.workspace = true
 
 [features]
 optimism = [
-    "reth-node-api/optimism",
     "reth-network/optimism",
     "reth-primitives/optimism",
     "reth-transaction-pool/optimism",
@@ -43,5 +42,5 @@ optimism = [
     "reth-provider/optimism",
     "reth-rpc-types-compat/optimism",
     "reth-rpc/optimism",
-    "reth-optimism-payload-builder/optimism"
+    "reth-optimism-payload-builder/optimism",
 ]
