+++ crates/rpc/rpc-api/Cargo.toml
@@ -24,4 +24,3 @@ serde_json.workspace = true
 
 [features]
 client = ["jsonrpsee/client", "jsonrpsee/async-client"]
-optimism = ["reth-node-api/optimism"]
