+++ crates/rpc/rpc-engine-api/Cargo.toml
@@ -49,8 +49,4 @@ reth-payload-builder = { workspace = true, features = ["test-utils"] }
 assert_matches.workspace = true
 
 [features]
-optimism = [
-    "reth-primitives/optimism",
-    "reth-rpc-types/optimism",
-    "reth-node-api/optimism",
-]
+optimism = ["reth-primitives/optimism", "reth-rpc-types/optimism"]
