+++ crates/storage/provider/Cargo.toml
@@ -61,8 +61,4 @@ rand.workspace = true
 
 [features]
 test-utils = ["alloy-rlp", "reth-db/test-utils"]
-optimism = [
-  "reth-primitives/optimism",
-  "reth-interfaces/optimism",
-  "reth-node-api/optimism"
-]
+optimism = ["reth-primitives/optimism", "reth-interfaces/optimism"]
